Move parse_text_field to fields.py
